@@ -2,7 +2,7 @@ package com.donbrody.customkeyboard
 
 import android.content.Intent
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.widget.Button
 import android.widget.EditText
 import com.donbrody.customkeyboard.components.keyboard.CustomKeyboardView
